requirements_dev.txt
Django<2
pyJWT
djangorestframework
django-rest-swagger

django-filter

tox
